# openapi/cci-backend.openapi.yaml
openapi: 3.0.3
info:
  title: CCI Backend API
  description: |
    เอกสารรวม endpoint สำหรับนำเข้า Apidog / Postman / Insomnia (Import → OpenAPI).
    พอร์ต default จาก `PORT` ใน `.env` (มักเป็น 3006).
    ส่วน Auth / บาง module อาจมี Guard — ตั้งค่า Bearer token ใน Apidog หลัง login.

    Masters (`/masters/*`, `/masters/products/*`) อยู่ในไฟล์ `cci-backend.masters.openapi.yaml`
    — นำเข้าคู่กันในโปรเจกต์ Apidog เดียวกัน หรือ merge paths ด้วยเครื่องมือ bundle OpenAPI
  version: 0.0.1
servers:
  - url: http://localhost:{port}
    variables:
      port:
        default: "3006"
...
security:
  - bearerAuth: []
paths:
...
  /products/{id}:
    get:
      tags: [Products]
      summary: Get product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
    patch:
      tags: [Products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        "200":
          description: OK
    delete:
      tags: [Products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
...
  /production-plans:
    get:
      tags: [Production plans]
      summary: List plans
      responses:
        "200":
          description: OK
    post:
      tags: [Production plans]
      summary: Create plan
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        "201":
          description: Created
...
  /production-plans/{id}/details:
    get:
      tags: [Production plans]
      summary: Plan details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
  /production-plans/{id}:
    get:
      tags: [Production plans]
      summary: Get plan
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
    patch:
      tags: [Production plans]
      summary: Update plan
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        "200":
          description: OK
    delete:
      tags: [Production plans]
      summary: Delete plan
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
  /production-plans/{id}/items:
    post:
      tags: [Production plans]
      summary: Add plan item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        "201":
          description: Created
  /production-plans/{id}/items/{itemId}:
    patch:
      tags: [Production plans]
      summary: Update plan item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: itemId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        "200":
          description: OK
    delete:
      tags: [Production plans]
      summary: Remove plan item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: itemId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
  /production-plans/{id}/reserve:
    post:
      tags: [Production plans]
      summary: Reserve materials
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
  /production-plans/{id}/generate-product-qr-orders:
    post:
      tags: [Production plans]
      summary: Generate production orders + QR lots (แผนต้อง confirmed)
      description: |
        จำนวน QR = ceil(quantity / lotSize).
        lotSize จาก defaultLotSize, lotSizeByProductId, product.lot_size หรือ 100
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                defaultLotSize: { type: number, example: 100 }
                planItemIds:
                  type: array
                  items: { type: integer }
                lotSizeByProductId:
                  type: object
                  additionalProperties: { type: number }
      responses:
        "200":
          description: OK
        "400":
          description: Bad request (e.g. not confirmed)
  /production-plans/{id}/confirm:
    post:
      tags: [Production plans]
      summary: Confirm plan
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
  /production-plans/{id}/confirm-and-issue:
    post:
      tags: [Production plans]
      summary: Confirm and issue
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
  /production-plans/{id}/issue:
    post:
      tags: [Production plans]
      summary: Issue materials
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
  /production-plans/{id}/cancel:
    post:
      tags: [Production plans]
      summary: Cancel plan
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
...
  /production-plans/debug/material/{materialCode}:
    get:
      tags: [Production plans]
      summary: Debug material
      parameters:
        - name: materialCode
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
  /production-plans/debug/check-availability/{materialId}:
    get:
      tags: [Production plans]
      summary: Check availability
      parameters:
        - name: materialId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: ใส่ token หลัง login (ถ้า route มี guard)